Fix issue with VS runtime checks for data truncation
When running VS with the 'c' runtime check on for data truncation (see https://docs.microsoft.com/en-us/cpp/build/reference/rtc-run-time-error-checks?view=msvc-160)

This code will throw errors because the 64 bit ptr gets truncated to a 32bit value. Normally I'd disable this using the pragma specifically for this code, but sadly that doesn't work for template functions, so I have come up with a way to get the appropriate masking type for any template combinations and using the recommended way to silence the error without impacting the quality of the generated code.

This code could be more concise in modern C++ but it seems the project expects to support C++03 so this should work fine there.

All tests passed locally for me on Windows (some asserts but verified they were the same before my changes)